# docs/contracts/agriinsight-analytics-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyticsEnvelope_CostsPayload_": {
        "additionalProperties": false,
        "properties": {
          "freshness": {
            "$ref": "#/components/schemas/FreshnessModel"
          },
          "lineage": {
            "$ref": "#/components/schemas/LineageModel"
          },
          "payload": {
            "$ref": "#/components/schemas/CostsPayload"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeModel"
          }
        },
        "required": [
          "freshness",
          "lineage",
          "payload",
          "scope"
        ],
        "title": "AnalyticsEnvelope[CostsPayload]",
        "type": "object"
      },
...
      "AnalyticsEnvelope_ProcurementCostsPayload_": {
        "additionalProperties": false,
        "properties": {
          "freshness": {
            "$ref": "#/components/schemas/FreshnessModel"
          },
          "lineage": {
            "$ref": "#/components/schemas/LineageModel"
          },
          "payload": {
            "$ref": "#/components/schemas/ProcurementCostsPayload"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeModel"
          }
        },
        "required": [
          "freshness",
          "lineage",
          "payload",
          "scope"
        ],
        "title": "AnalyticsEnvelope[ProcurementCostsPayload]",
        "type": "object"
      },
...
      "CostBreakdownModel": {
        "additionalProperties": false,
        "properties": {
          "activityType": {
            "title": "Activitytype",
            "type": "string"
          },
          "laborCostVnd": {
            "title": "Laborcostvnd",
            "type": "number"
          },
          "materialCostVnd": {
            "title": "Materialcostvnd",
            "type": "number"
          },
          "sharePct": {
            "title": "Sharepct",
            "type": "number"
          },
          "totalCostVnd": {
            "title": "Totalcostvnd",
            "type": "number"
          }
        },
        "required": [
          "activityType",
          "materialCostVnd",
          "laborCostVnd",
          "totalCostVnd",
          "sharePct"
        ],
        "title": "CostBreakdownModel",
        "type": "object"
      },
      "CostCapabilitiesModel": {
        "additionalProperties": false,
        "properties": {
          "activityBreakdownAvailable": {
            "title": "Activitybreakdownavailable",
            "type": "boolean"
          },
          "fileExportAvailable": {
            "const": false,
            "title": "Fileexportavailable",
            "type": "boolean"
          },
          "monthlyBreakdownAvailable": {
            "title": "Monthlybreakdownavailable",
            "type": "boolean"
          },
          "readOnly": {
            "const": true,
            "title": "Readonly",
            "type": "boolean"
          }
        },
        "required": [
          "readOnly",
          "fileExportAvailable",
          "monthlyBreakdownAvailable",
          "activityBreakdownAvailable"
        ],
        "title": "CostCapabilitiesModel",
        "type": "object"
      },
      "CostFarmModel": {
        "additionalProperties": false,
        "properties": {
          "budgetOperatingCostVnd": {
            "title": "Budgetoperatingcostvnd",
            "type": "number"
          },
          "budgetVarianceVnd": {
            "title": "Budgetvariancevnd",
            "type": "number"
          },
          "farmCode": {
            "title": "Farmcode",
            "type": "string"
          },
          "farmName": {
            "title": "Farmname",
            "type": "string"
          },
          "fieldCount": {
            "title": "Fieldcount",
            "type": "integer"
          },
          "harvestQuantityKg": {
            "title": "Harvestquantitykg",
            "type": "number"
          },
          "operatingCostPerHaVnd": {
            "title": "Operatingcostperhavnd",
            "type": "number"
          },
          "operatingCostPerKgVnd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operatingcostperkgvnd"
          },
          "operatingLaborCostVnd": {
            "title": "Operatinglaborcostvnd",
            "type": "number"
          },
          "operatingMaterialCostVnd": {
            "title": "Operatingmaterialcostvnd",
            "type": "number"
          },
          "operatingProfitMarginPct": {
            "title": "Operatingprofitmarginpct",
            "type": "number"
          },
          "operatingProfitVnd": {
            "title": "Operatingprofitvnd",
            "type": "number"
          },
          "operatingTotalCostVnd": {
            "title": "Operatingtotalcostvnd",
            "type": "number"
          },
          "province": {
            "title": "Province",
            "type": "string"
          },
          "revenueVnd": {
            "title": "Revenuevnd",
            "type": "number"
          },
          "seasonAreaHa": {
            "title": "Seasonareaha",
            "type": "number"
          },
          "seasonCount": {
            "title": "Seasoncount",
            "type": "integer"
          },
          "targetYieldKg": {
            "title": "Targetyieldkg",
            "type": "number"
          }
        },
        "required": [
          "farmCode",
          "farmName",
          "province",
          "seasonCount",
          "fieldCount",
          "seasonAreaHa",
          "budgetOperatingCostVnd",
          "targetYieldKg",
          "harvestQuantityKg",
          "revenueVnd",
          "operatingMaterialCostVnd",
          "operatingLaborCostVnd",
          "operatingTotalCostVnd",
          "operatingProfitVnd",
          "operatingProfitMarginPct",
          "operatingCostPerHaVnd",
          "operatingCostPerKgVnd",
          "budgetVarianceVnd"
        ],
        "title": "CostFarmModel",
        "type": "object"
      },
      "CostMonthlyModel": {
        "additionalProperties": false,
        "properties": {
          "month": {
            "title": "Month",
            "type": "string"
          },
          "operatingLaborCostVnd": {
            "title": "Operatinglaborcostvnd",
            "type": "number"
          },
          "operatingMaterialCostVnd": {
            "title": "Operatingmaterialcostvnd",
            "type": "number"
          },
          "operatingProfitMarginPct": {
            "title": "Operatingprofitmarginpct",
            "type": "number"
          },
          "operatingProfitVnd": {
            "title": "Operatingprofitvnd",
            "type": "number"
          },
          "operatingTotalCostVnd": {
            "title": "Operatingtotalcostvnd",
            "type": "number"
          },
          "revenueVnd": {
            "title": "Revenuevnd",
            "type": "number"
          }
        },
        "required": [
          "month",
          "operatingMaterialCostVnd",
          "operatingLaborCostVnd",
          "operatingTotalCostVnd",
          "revenueVnd",
          "operatingProfitVnd",
          "operatingProfitMarginPct"
        ],
        "title": "CostMonthlyModel",
        "type": "object"
      },
...
      "CostsPayload": {
        "additionalProperties": false,
        "properties": {
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/CostBreakdownModel"
            },
            "title": "Breakdown",
            "type": "array"
          },
          "capabilities": {
            "$ref": "#/components/schemas/CostCapabilitiesModel"
          },
          "farms": {
            "items": {
              "$ref": "#/components/schemas/CostFarmModel"
            },
            "title": "Farms",
            "type": "array"
          },
          "monthly": {
            "items": {
              "$ref": "#/components/schemas/CostMonthlyModel"
            },
            "title": "Monthly",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/CostSummaryModel"
          }
        },
        "required": [
          "breakdown",
          "capabilities",
          "farms",
          "monthly",
          "summary"
        ],
        "title": "CostsPayload",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "correlationId": {
            "title": "Correlationid",
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "correlationId",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FreshnessModel": {
        "additionalProperties": false,
        "properties": {
          "artifactAgeHours": {
            "minimum": 0.0,
            "title": "Artifactagehours",
            "type": "number"
          },
          "dataStatus": {
            "enum": [
              "current",
              "stale",
              "partial",
              "missing"
            ],
            "title": "Datastatus",
            "type": "string"
          },
          "maxAgeHours": {
            "minimum": 1.0,
            "title": "Maxagehours",
            "type": "integer"
          }
        },
        "required": [
          "artifactAgeHours",
          "dataStatus",
          "maxAgeHours"
        ],
        "title": "FreshnessModel",
        "type": "object"
      },
...
      "LineageModel": {
        "additionalProperties": false,
        "properties": {
          "asOf": {
            "title": "Asof",
            "type": "string"
          },
          "contractVersion": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Contractversion",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "manifestFingerprint": {
            "title": "Manifestfingerprint",
            "type": "string"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          }
        },
        "required": [
          "asOf",
          "generatedAt",
          "manifestFingerprint",
          "runId"
        ],
        "title": "LineageModel",
        "type": "object"
      },
...
      "ProcurementCostCapabilitiesModel": {
        "additionalProperties": false,
        "properties": {
          "detailPageAvailable": {
            "const": true,
            "title": "Detailpageavailable",
            "type": "boolean"
          },
          "fileExportAvailable": {
            "const": true,
            "title": "Fileexportavailable",
            "type": "boolean"
          },
          "readOnly": {
            "const": true,
            "title": "Readonly",
            "type": "boolean"
          }
        },
        "required": [
          "readOnly",
          "fileExportAvailable",
          "detailPageAvailable"
        ],
        "title": "ProcurementCostCapabilitiesModel",
        "type": "object"
      },
...
      "ProcurementCostsPayload": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/ProcurementCostCapabilitiesModel"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcurementCostItemModel"
            },
            "title": "Items",
            "type": "array"
          },
          "monthly": {
            "items": {
              "$ref": "#/components/schemas/ProcurementCostMonthlyModel"
            },
            "title": "Monthly",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/ProcurementPageModel"
          },
          "summary": {
            "$ref": "#/components/schemas/ProcurementCostSummaryModel"
          },
          "suppliers": {
            "items": {
              "$ref": "#/components/schemas/ProcurementCostSupplierModel"
            },
            "title": "Suppliers",
            "type": "array"
          }
        },
        "required": [
          "capabilities",
          "items",
          "monthly",
          "page",
          "suppliers",
          "summary"
        ],
        "title": "ProcurementCostsPayload",
        "type": "object"
      },
...
      "ScopeModel": {
        "additionalProperties": false,
        "properties": {
          "appliedFilter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppliedFilterModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "farmCodes": {
            "items": {
              "type": "string"
            },
            "title": "Farmcodes",
            "type": "array"
          },
          "tenantId": {
            "title": "Tenantid",
            "type": "string"
          },
          "tenantWide": {
            "title": "Tenantwide",
            "type": "boolean"
          },
          "warehouseCodes": {
            "items": {
              "type": "string"
            },
            "title": "Warehousecodes",
            "type": "array"
          }
        },
        "required": [
          "tenantId",
          "tenantWide"
        ],
        "title": "ScopeModel",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only, demo-tenant-gated analytics over checksum-verified aggregate snapshots.",
    "title": "AgriInsight Internal Analytics API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/costs": {
      "get": {
        "operationId": "getAnalyticsCosts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsEnvelope_CostsPayload_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Costs",
        "tags": [
          "analytics-costs"
        ]
      }
    },
    "/internal/v1/costs/export": {
      "get": {
        "operationId": "getAnalyticsCostExport",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": true,
            "schema": {
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "all",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "farm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm"
            }
          },
          {
            "in": "query",
            "name": "crop",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crop"
            }
          },
          {
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activity"
            }
          },
          {
            "in": "query",
            "name": "supplier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier"
            }
          },
          {
            "in": "query",
            "name": "season",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Season"
            }
          },
          {
            "in": "query",
            "name": "month_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month From"
            }
          },
          {
            "in": "query",
            "name": "month_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month To"
            }
          },
          {
            "in": "query",
            "name": "top_n",
            "required": false,
            "schema": {
              "default": 15,
              "maximum": 30,
              "minimum": 1,
              "title": "Top N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Cost Export",
        "tags": [
          "analytics-cost-exports"
        ]
      }
    },
    "/internal/v1/costs/procurement": {
      "get": {
        "operationId": "getProcurementCosts",
        "parameters": [
          {
            "in": "query",
            "name": "farm_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm Code"
            }
          },
          {
            "in": "query",
            "name": "month_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month From"
            }
          },
          {
            "in": "query",
            "name": "month_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 10000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsEnvelope_ProcurementCostsPayload_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Procurement Costs",
        "tags": [
          "analytics-costs"
        ]
      }
    },
...
  }
}